refactor(ptt): extract client session dispatch
Why:
- Fixes #78 by giving Daemon ServerMessage dispatch a named Client Session Module instead of keeping the policy buried in app.rs.
- The extraction improves Locality for Session, Stream path, Seal path, Overlay, Injection, and Runtime truth handling without changing the Client runtime behavior.
- Keeping test helpers local is intentional for this slice; #79 owns the shared fixture cleanup once this Module boundary is stable.

What changed:
- Moved ServerMessage dispatch policy, parent focus enqueue handling, Session state helpers, timing labels, and error classification into parakeet-ptt/src/client_session.rs.
- Moved the dispatch behavior tests out of app.rs into the new Module test surface.
- Kept app.rs as the runtime orchestrator and main.rs as the binary Adapter that wires the new Module.

Compatibility:
- No user-facing behavior change intended.
- Final results still enqueue exactly one Injection job without waiting for completion.
- Interim Stream path messages still route to Overlay without enqueueing Injection.
- Session ended/error paths still clear transient Session state and parent focus data.
- Overlay failure remains non-fatal to final Injection.

Verification (#78):
- targeted: cargo test client_session --manifest-path Cargo.toml (from parakeet-ptt/) -> PASSED
- targeted: cargo test app_hotkey_press_sends_start_session_message --manifest-path Cargo.toml (from parakeet-ptt/) -> PASSED
- targeted: cargo test app_injection_report_error_routes_failure_classification --manifest-path Cargo.toml (from parakeet-ptt/) -> PASSED
- regression: N/A, refactor with existing behavior coverage moved into client_session.rs
- full suite: cargo test --manifest-path parakeet-ptt/Cargo.toml -> PASSED
- lint: cargo clippy --manifest-path parakeet-ptt/Cargo.toml --all-targets --all-features -- -D warnings -> PASSED
- types: cargo check --manifest-path parakeet-ptt/Cargo.toml --all-targets --all-features -> PASSED
- dead-code: N/A, no repo dead-code gate for Rust discovered
- build: cargo build --manifest-path parakeet-ptt/Cargo.toml --bins -> PASSED
- format: cargo fmt --manifest-path parakeet-ptt/Cargo.toml -- --check -> PASSED
- pre-commit: prek run --files parakeet-ptt/src/app.rs parakeet-ptt/src/main.rs parakeet-ptt/src/client_session.rs -> PASSED
- pre-commit: prek run --stage pre-push --files parakeet-ptt/src/app.rs parakeet-ptt/src/main.rs parakeet-ptt/src/client_session.rs -> PASSED
- pre-commit: prek run --all-files -> PASSED
- pre-commit: prek run --stage pre-push --all-files -> PASSED
- static/security: N/A, no touched security/migration/workflow surface
- migrations: N/A
- CLI/script smoke: N/A
- UI render: N/A
- destructive/negative: N/A
- review: coderabbit_review_watch.py local --fallback-codex-review-args "--base main" -- --base main -> clean, CodeRabbit backend, .git/coderabbit-review/20260519T115639Z/status.json
- PR gate: pending until PR exists
- tests added/expanded: existing Client dispatch behavior tests moved from app.rs to client_session.rs; no new behavior branch added
- preexisting failures left: none

Independent review:
- Local CodeRabbit review reported 0 findings. No MUST_FIX, DEFER, or IGNORE items remain from local review.

Deferred follow-ups:
- None for this issue. Shared Client runtime test fixtures remain owned by #79.
